package restructure, code tidy up contd..
