+++ internal/openai/client.go
