+++ internal/pinecone/config.go
