+++ internal/pinecone/pinecone_adapter.go
