+++ internal/pinecone/pinecone_test.go
